RFC: Plugins 1.1
Stabilize a very small sliver of today's plugin system in the compiler, just
enough to get basic features like custom derive working. Ensure that the
features stabilized here will not pose a maintenance burden on the compiler but
also don't try to stabilize enough features for the "perfect macro system" at
the same time. Overall, this should be considered an incremental step towards an
official "plugins 2.0".